requirements.txt
absl-py==2.3.1
altgraph==0.17.4
asttokens==3.0.0
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
cmake==4.1.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
dlib @ file:///C:/Users/Alhousainy/Downloads/dlib-19.22.99-cp310-cp310-win_amd64.whl#sha256=f2181b6724669fb0147c9ebc326d72b94de41bced1812eff5df4b133b5d0b575
exceptiongroup==1.3.0
executing==2.2.1
flatbuffers==25.2.10
fonttools==4.59.2
gast==0.6.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==2.15.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
MarkupSafe==3.0.2
matplotlib==3.7.3
matplotlib-inline==0.1.7
ml-dtypes==0.2.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.3.1
opencv-python==4.9.0.80
opt_einsum==3.4.0
packaging==25.0
parso==0.8.5
pefile==2023.2.7
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==4.25.8
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pywin32==311
pywin32-ctypes==0.2.3
pyzmq==27.0.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.2.2
scipy==1.15.3
six==1.17.0
stack-data==0.6.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.14.2
